Refactor: 파일 이름 수정 및 DTO 객체 record 클래스로 변환 (#36)
- MemberNicknameServiceRequest 클래스를 불변 클래스인 record 클래스로 변환했습니다.
- MemberNicknameRequest 클래스의 필드 변수가 1개이므로 Builder 패턴을 제거하고, 생성자 형태로 수정하여 가독성을 높였습니다.
- CommonConstant 이름이 디렉터리 구조와 맞지 않다고 생각하여 ConstantUtil 파일명으로 수정했습니다.
- 그 외 파일들의 변화는 파일 이름 변화로 인한 수정 내용입니다.

issue #36
