+++ src/main/java/dasi/typing/jwt/JwtAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package dasi.typing.jwt;
 
-import static dasi.typing.utils.CommonConstant.BEARER_PREFIX;
-import static dasi.typing.utils.CommonConstant.REISSUE_URI;
-import static dasi.typing.utils.CommonConstant.TOKEN_HEADER;
+import static dasi.typing.utils.ConstantUtil.BEARER_PREFIX;
+import static dasi.typing.utils.ConstantUtil.REISSUE_URI;
+import static dasi.typing.utils.ConstantUtil.TOKEN_HEADER;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
